+++ release_notes.md
@@ -112,3 +112,6 @@ First release into griffon artifacts portal.
 1. New nested attributes support in findByDsl() by separating attributes using double underscores(__), for example: item__category__name eq('CAT1').
 1. Fixes create-simple-jpa failed with error 'java.sql.SQLException: No suitable driver found'.
 1. Add '-jdbc=derby-embedded' to create-simple-jpa command.
+
+### 0.5.3
+1. Allow user to specify default content decorator for `mvcPopupButton()` in `DialogUtils`.